Extend interface token and improved DHCPv6 PD route installation.
Having to specify the execute statements in all subnet section and
per interface was a burden. I extended the 'interface' such that it
can be used in an 'execute' statement in case a packet is received.
I also improved the route installation/removal  script such that it
updates routes cleanly (using ip add, del and change) only.